# docs/openapi-v1.yaml
openapi: 3.1.0
info:
  title: LQEpoch Market Data Platform Read API
  version: 1.0.0
  description: >-
    Read-only authenticated access to verified MDP datasets. Bars are the existing
    lqepoch.us_equity_trade_bar_1m.v1 contract. They do not carry DatasetManifestV2
    completion evidence. A future BarV2 response must use a separately versioned contract. The
    optional Linux LocalTest Pair route verifies one exact existing chunk and does not establish a
    capture rollup or provider completeness. It is mounted only when explicit LocalTest state and
    staging roots are configured at startup. The
    router adds Cache-Control: no-store to every response, including authorization, extractor,
    fallback, and overload errors; an existing Cache-Control value is preserved.
servers:
  - url: http://127.0.0.1:8088
    description: Default loopback-only listener
paths:
  /healthz:
    get:
      operationId: liveness
      summary: Process liveness
      responses:
        "200":
          description: The HTTP process is serving requests.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    const: alive
  /readyz:
    get:
      operationId: readiness
      summary: Identity and transport configuration status
      responses:
        "200":
          description: Required independent MDP verifier keys are configured.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Readiness"
        "503":
          description: The service is restricted to liveness because MDP keys are not configured.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Readiness"
  /v1/datasets/{dataset_id}/bars:
    get:
      operationId: getDatasetBarsV1
      summary: Read verified one-minute trade bars
      description: >-
        The JWT is checked before any request-level storage access or query. LocalTest transport
        serves only the diagnostic namespace and always reports synthetic/synthetic/unknown
        provenance. Curated reads require an independently authorized archive. No capture, upload,
        account, or order operation is exposed by this bars route. A separate optional LocalTest
        chunk-verification route does not publish or ingest data.
      security:
        - mdpDelegation: []
      parameters:
        - name: dataset_id
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 128
        - name: namespace
          in: query
          required: true
          schema:
            type: string
            enum: [diagnostic, curated]
        - name: symbol
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 16
            pattern: "^[A-Z0-9.-]+$"
      responses:
        "200":
          description: Verified V1 bars and the existing remote query summary.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BarsResponseV1"
        "400":
          description: Invalid symbol, dataset identifier, or query parameters.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid market:read delegation.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Requested namespace is not authorized for the configured transport.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: The bounded result limit was exceeded.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "504":
          description: Query deadline expired; in-flight worker cancellation is supervised.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: No authenticated service or verified query result is available.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v2/local-test/capture-pairs/{receipt_sha256}/verify:
    get:
      operationId: verifyOneLocalTestCapturePairChunkV2
      summary: Verify one explicitly named LocalTest Pair chunk
      description: >-
        This Linux-only route is mounted only when the service starts with an existing LocalTest
        root plus explicit Pair state and staging roots. Authentication is checked before the
        receipt or archive is accessed. The canonical receipt basename is constructed from the
        lowercase SHA-256 path value; request paths and receipt directory scans are never accepted.
        The bounded worker verifies exactly one receipt, both manifests and objects, and the Core
        raw/event join. The compact response contains no rows, raw payloads, or local paths. It
        preserves one validated provider/feed pair (alpaca/opra or synthetic/synthetic) and always reports
        source_completeness NOT_ASSERTED. A verified chunk is not a complete capture, provider
        entitlement, freshness, Drive durability, or historical batch.
      security:
        - mdpDelegation: []
      parameters:
        - name: receipt_sha256
          in: path
          required: true
          description: Lowercase hexadecimal SHA-256 of one existing canonical chunk receipt.
          schema:
            type: string
            minLength: 64
            maxLength: 64
            pattern: "^[0-9a-f]{64}$"
      responses:
        "200":
          description: Exact readback verification of one local chunk only.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LocalCapturePairChunkSummaryV2"
        "400":
          description: Receipt hash is not canonical lowercase SHA-256.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid market:read delegation.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Pair verification is not authorized for the configured transport.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: Receipt or artifacts exceed the bounded verifier limits.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Pair verification failed, storage lock is busy, or the service is at capacity.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "504":
          description: Request deadline expired and owned worker cancellation is supervised.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  headers:
    NoStore:
      description: Response is not cacheable by HTTP caches.
      schema:
        type: string
        const: no-store
  securitySchemes:
    mdpDelegation:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        HS256 only; audience lqepoch-market-data; exact scope [market:read]; maximum lifetime
        60 seconds; issuer and key are selected by kid: mdp-terminal maps to
        eqoboard-openterminal and MDP_TERMINAL_JWT_SECRET, while mdp-research maps to
        openterminal-research and MDP_RESEARCH_JWT_SECRET. These keys are distinct from each
        other and from Gateway keys. Each BFF signer receives only its matching MDP key; the MDP
        verifier receives both.
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          enum:
            - invalid_request
            - unauthorized
            - not_authorized
            - result_limit_exceeded
            - query_timeout
            - query_unavailable
    Readiness:
      type: object
      required: [status, market_ready, source_entitlement]
      properties:
        status:
          type: string
          enum: [ready, not_ready]
        market_ready:
          const: false
          description: Remains false until trusted source and entitlement evidence is integrated.
        source_entitlement:
          const: unverified
    BarsResponseV1:
      type: object
      required: [summary, rows]
      properties:
        summary:
          $ref: "#/components/schemas/RemoteQuerySummary"
        rows:
          type: array
          maxItems: 390
          items:
            $ref: "#/components/schemas/TradeMinuteBarV1"
    LocalCapturePairChunkSummaryV2:
      type: object
      additionalProperties: false
      required:
        - status
        - verification_scope
        - source_completeness
        - pair_verification
        - pair_verification_version
        - transport
        - provider
        - feed
        - entitlement
        - capture_instance_id
        - source_generation
        - canonical_generation
        - first_source_frame_sequence
        - last_source_frame_sequence
        - raw_frame_count
        - normalized_event_count
        - input_payload_bytes
        - input_chunk_sha256
        - manifest_input_sha256
        - raw_dataset_id
        - raw_manifest_sha256
        - raw_object_sha256
        - raw_schema_id
        - raw_row_count
        - event_dataset_id
        - event_manifest_sha256
        - event_object_sha256
        - event_schema_id
        - event_row_count
      properties:
        status: {type: string, const: VERIFIED_LOCAL_TEST_CHUNK_ONLY}
        verification_scope: {type: string, const: SINGLE_CHUNK_LOCAL_READBACK}
        source_completeness: {type: string, const: NOT_ASSERTED}
        pair_verification:
          type: string
          const: EXACT_RAW_EVENT_ARTIFACT_READBACK_MATCH_LOCAL_ONLY
        pair_verification_version: {type: integer, const: 1}
        transport: {type: string, const: local_test}
        provider: {type: string, enum: [alpaca, synthetic]}
        feed: {type: string, enum: [opra, synthetic]}
        entitlement: {type: string, const: unknown}
        capture_instance_id:
          type: string
          pattern: "^[0-9a-f]{12}4[0-9a-f]{3}[89ab][0-9a-f]{15}$"
        source_generation: {$ref: "#/components/schemas/WireU64"}
        canonical_generation: {$ref: "#/components/schemas/WireU64"}
        first_source_frame_sequence: {$ref: "#/components/schemas/WireU64"}
        last_source_frame_sequence: {$ref: "#/components/schemas/WireU64"}
        raw_frame_count:
          type: integer
          minimum: 1
          maximum: 1024
          description: Exact raw frame rows in this chunk; raw_row_count has the same value.
        normalized_event_count:
          type: integer
          minimum: 0
          maximum: 524288
          description: Exact normalized event rows in this chunk; event_row_count has the same value.
        input_payload_bytes:
          allOf:
            - $ref: "#/components/schemas/WireU64"
          description: Sum of the exact ordered captured raw payload byte lengths.
        input_chunk_sha256:
          type: string
          pattern: "^[0-9a-f]{64}$"
          description: Metadata-rich MDP chunk identity digest, distinct from the manifest payload-only input hash.
        manifest_input_sha256:
          type: string
          pattern: "^[0-9a-f]{64}$"
          description: Core finite-batch SHA-256 of the exact ordered raw payload byte concatenation.
        raw_dataset_id: {type: string}
        raw_manifest_sha256: {type: string, pattern: "^[0-9a-f]{64}$"}
        raw_object_sha256: {type: string, pattern: "^[0-9a-f]{64}$"}
        raw_schema_id: {type: string, enum: [lqepoch.market_raw_frame.v2, lqepoch.market_raw_json_frame.v2]}
        raw_row_count: {$ref: "#/components/schemas/WireU64"}
        event_dataset_id: {type: string}
        event_manifest_sha256: {type: string, pattern: "^[0-9a-f]{64}$"}
        event_object_sha256: {type: string, pattern: "^[0-9a-f]{64}$"}
        event_schema_id: {type: string, const: lqepoch.market_event.v3}
        event_row_count: {$ref: "#/components/schemas/WireU64"}
      allOf:
        - oneOf:
            - properties:
                provider: {const: alpaca}
                feed: {const: opra}
            - properties:
                provider: {const: synthetic}
                feed: {const: synthetic}
    RemoteQuerySummary:
      type: object
      required:
        - namespace
        - dataset_id
        - schema_id
        - source
        - row_count
        - returned_rows
        - content_sha256
        - parquet_schema_sha256
        - cache_hit
      properties:
        namespace:
          type: string
          enum: [diagnostic, curated]
        dataset_id:
          type: string
        schema_id:
          const: lqepoch.us_equity_trade_bar_1m.v1
        source:
          $ref: "#/components/schemas/MarketDataSourceV1"
        row_count:
          description: Canonical base-10 string projection of an unsigned 64-bit value.
          $ref: "#/components/schemas/WireU64"
        returned_rows:
          description: Canonical base-10 string projection of an unsigned 64-bit value.
          $ref: "#/components/schemas/WireU64"
        content_sha256:
          type: string
          pattern: "^[0-9a-f]{64}$"
        parquet_schema_sha256:
          type: string
          pattern: "^[0-9a-f]{64}$"
        cache_hit:
          type: boolean
    MarketDataSourceV1:
      type: object
      required: [provider, feed, entitlement, numeric_encoding]
      properties:
        provider:
          type: string
        feed:
          type: string
        entitlement:
          type: string
          enum: [unknown, authorized, unauthorized]
        numeric_encoding:
          type: string
          enum:
            - unspecified
            - decimal_token
            - integer_token
            - binary_float64_shortest_decimal
            - binary_float32_shortest_decimal
            - raw_messagepack_bytes
            - raw_json_bytes
        source_record_id:
          type: [string, "null"]
    TradeMinuteBarV1:
      type: object
      additionalProperties: false
      description: >-
        One observed trade-minute bar. The window may be sparse: minutes without a trade observation,
        including quote-only minutes, have no row. For each symbol, emitted row count equals
        window_expected_minutes minus window_empty_trade_minutes. The count describes the consumed
        finite input and does not prove provider-side market completeness or entitlement.
      required:
        - schema_version
        - source_provider
        - source_feed
        - source_entitlement
        - source_numeric_encoding
        - symbol
        - bar_start_utc
        - bar_end_exclusive_utc
        - available_at_utc
        - trade_date
        - session_id
        - session_timezone
        - session_policy_id
        - session_policy_sha256
        - session_start_utc
        - session_end_exclusive_utc
        - window_start_utc
        - window_end_exclusive_utc
        - open
        - high
        - low
        - close
        - volume
        - trade_count
        - quote_events_excluded
        - source_timestamp_missing_rows
        - sequence_gap_count
        - late_event_count
        - window_expected_minutes
        - window_empty_trade_minutes
        - source_start_utc
        - source_end_exclusive_utc
        - window_input_eof
        - source_pages_exhausted
        - completion_mode
        - nbbo_input_status
      properties:
        schema_version:
          type: integer
          const: 1
        source_provider: {type: string}
        source_feed: {type: string}
        source_entitlement: {type: string}
        source_numeric_encoding: {type: string}
        symbol: {type: string}
        bar_start_utc: {$ref: "#/components/schemas/UtcTimestamp"}
        bar_end_exclusive_utc: {$ref: "#/components/schemas/UtcTimestamp"}
        available_at_utc: {$ref: "#/components/schemas/UtcTimestamp"}
        trade_date: {type: string, format: date}
        session_id: {type: string}
        session_timezone: {type: string}
        session_policy_id: {type: string}
        session_policy_sha256: {type: string, pattern: "^[0-9a-f]{64}$"}
        session_start_utc: {$ref: "#/components/schemas/UtcTimestamp"}
        session_end_exclusive_utc: {$ref: "#/components/schemas/UtcTimestamp"}
        window_start_utc: {$ref: "#/components/schemas/UtcTimestamp"}
        window_end_exclusive_utc: {$ref: "#/components/schemas/UtcTimestamp"}
        open: {$ref: "#/components/schemas/ExactDecimalString"}
        high: {$ref: "#/components/schemas/ExactDecimalString"}
        low: {$ref: "#/components/schemas/ExactDecimalString"}
        close: {$ref: "#/components/schemas/ExactDecimalString"}
        volume: {$ref: "#/components/schemas/ExactDecimalString"}
        trade_count: {$ref: "#/components/schemas/WireU64"}
        quote_events_excluded: {$ref: "#/components/schemas/WireU64"}
        source_timestamp_missing_rows: {$ref: "#/components/schemas/WireU64"}
        sequence_gap_count: {$ref: "#/components/schemas/WireU64"}
        late_event_count: {$ref: "#/components/schemas/WireU64"}
        window_expected_minutes: {$ref: "#/components/schemas/WireU64"}
        window_empty_trade_minutes: {$ref: "#/components/schemas/WireU64"}
        source_start_utc: {$ref: "#/components/schemas/UtcTimestamp"}
        source_end_exclusive_utc: {$ref: "#/components/schemas/UtcTimestamp"}
        window_input_eof: {type: boolean}
        source_pages_exhausted:
          type: [boolean, "null"]
        completion_mode: {type: string}
        nbbo_input_status: {type: string}
    UtcTimestamp:
      type: string
      format: date-time
      pattern: "Z$"
    ExactDecimalString:
      type: string
      pattern: "^-?(0|[1-9][0-9]*)(\\.[0-9]+)?$"
    WireU64:
      type: string
      pattern: "^(0|[1-9][0-9]*)$"